implemented crud operation for user role (ui, server)

+++ gwt/common/src/main/java/ru/protei/portal/ui/common/client/events/RoleEvents.java
@@ -55,10 +55,10 @@ public class RoleEvents {
     @Url( value = "role" )
     public static class Edit {
 
-        public Integer id;
+        public Long id;
 
         public Edit() {}
 
-        public Edit( Integer id ) { this.id = id; }
+        public Edit( Long id ) { this.id = id; }
     }
 }
